fix: resolve dtype mismatch in compute_layer_entropy_logit_lens
Cast hidden state to lm_head.weight.dtype instead of hardcoding .float()
(float32). When the model is loaded in float16 or bfloat16, the norm layer
and lm_head weights are half-precision, causing a RuntimeError on the
matrix multiply. Using .to(lm_head.weight.dtype) makes the cast dynamic
and correct for float16, bfloat16, and float32 models alike.

Fixes: RuntimeError: expected mat1 and mat2 to have the same dtype,
but got: float != c10::Half

+++ src/evaluation/acc_adaptive_evaluator_v5.py
@@ -166,7 +166,7 @@ def compute_layer_entropy_logit_lens(
         if key not in hidden_states:
             entropies.append(0.0)
             continue
-        h = hidden_states[key].to(lm_head.weight.device).float()
+        h = hidden_states[key].to(lm_head.weight.device).to(lm_head.weight.dtype)
         if hasattr(model, 'model') and hasattr(model.model, 'norm'):
             h = model.model.norm(h.unsqueeze(0)).squeeze(0)
         with torch.no_grad():
